fix: export useful constants
