+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "arweave-stream-bundle",
-  "version": "0.2.2",
+  "version": "0.3.0",
   "license": "MIT",
   "keywords": [
     "arweave",
