+++ src/models/data-item-header.ts
@@ -10,12 +10,30 @@ import {
   readByteArrayFromByteReader,
   readOptionalByteArrayFromByteReader,
   readUintLEFromByteReader,
-  SignatureType,
-  SIGNATURE_TYPE_CONSTANTS,
   Tag,
   uintLEToByteArray,
 } from '../utils';
 import { DeserializationResult } from './deserialization-result';
+import { SignatureType } from './signature-type';
+
+const SIGNATURE_TYPE_CONSTANTS: Record<SignatureType, { signatureByteLength: number; publicKeyByteLength: number }> = {
+  [SignatureType.PS256_65537]: {
+    signatureByteLength: 512,
+    publicKeyByteLength: 512,
+  },
+  [SignatureType.Ed25519]: {
+    signatureByteLength: 64,
+    publicKeyByteLength: 32,
+  },
+  [SignatureType.ES256K]: {
+    signatureByteLength: 65,
+    publicKeyByteLength: 65,
+  },
+  [SignatureType.ES256K_Compact]: {
+    signatureByteLength: 65,
+    publicKeyByteLength: 33,
+  },
+};
 
 export class DataItemHeader {
   id: string | null;
