+++ src/models/index.ts
@@ -1,4 +1,6 @@
 export * from './data-bundle-header';
 export * from './data-item';
 export * from './data-item-header';
+export * from './signature-type';
 export * from './tags';
+
