+++ src/models/signature-type.ts
@@ -1,0 +1,6 @@
+export enum SignatureType {
+  PS256_65537 = 1,
+  Ed25519 = 2,
+  ES256K = 3,
+  ES256K_Compact = 4,
+}
