+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './byte-codec';
-export * from './constants';
 export * from './deep-hash-stream';
 export * from './tag-codec';
